dynamic loading of directory open button based on browser
